refactor(admin): improve inventory table performance and product details drawer UX
